Extract LutronDevice into separate file
