+++ homeassistant/components/lutron/light.py
@@ -9,7 +9,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import LUTRON_CONTROLLER, LUTRON_DEVICES, LutronDevice
+from . import LUTRON_CONTROLLER, LUTRON_DEVICES
+from .entity import LutronDevice
 
 
 async def async_setup_entry(
